Makefile: provide an option to not build manpages
The process of generating the vpnc.8 man page consists in running the
vpnc tool itself, and parse its --long-help output. While this is
perfectly fine when building natively, it fails completely when
cross-compiling: the vpnc binary that was built cannot be executed on
the build machine.

In order to support such situations with minimal changes, this patch
adjusts the Makefile to make it understand a MANS variable. By
default, it's defined to "vpnc.8", which means the manpage continues
to be built as usual. However, if it's overriden to be empty, then no
manpage is built.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>