Makefile: allow to override the version
The mk-version script makes the assumption that the build process runs
from a SVN checkout or a Git clone of the vpnc source code. However,
this is not always the case, for example when you are building from a
tarball, but inside a larger project that is versioned with Git.

This is typically the case when building vpnc under a tool like
Buildroot, or potentially other embedded Linux build system: the build
system tool itself is versioned in Git, downloads tarballs of software
components and builds them.

In order to support such situations, this commit adjusts the Makefile
to allow the VERSION variable to be overriden from the make command
line.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>